reuse `ZoneEvmConfig`
